+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
-- Upcoming changes.
+-
+
+## [0.2.0] - 2025-07-03
+- Fixed PORT expansion in Dockerfile
+- Added database health endpoint and service info route
+- Documentation and badges updated
 
 ## [0.1.0] - 2025-05-22
 - Initial release.